Extracted RAC router configuration from router.tsx
